climbing: layers refactoring
